redid gruntfile and tasks to be more module
