+++ bower.json
@@ -4,7 +4,7 @@
   "homepage": "https://github.com/azinasili/yoga",
   "author": "Azin Asili @azinasili",
   "copyright": "2015",
-  "version": "0.2.5",
+  "version": "0.2.6",
   "license": "MIT",
   "main": "_yoga.scss",
   "keywords": [
