+++ package.json
@@ -4,7 +4,7 @@
   "homepage": "https://github.com/azinasili/yoga",
   "author": "Azin Asili @azinasili",
   "copyright": "2015",
-  "version": "0.2.5",
+  "version": "0.2.6",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -25,10 +25,9 @@
     "grunt": "^0.4.5",
     "grunt-autoprefixer": "^3.0.3",
     "grunt-contrib-connect": "^0.10.1",
-    "grunt-contrib-cssmin": "^0.13.0",
     "grunt-contrib-sass": "^0.9.2",
     "grunt-contrib-watch": "^0.6.1",
     "grunt-open": "^0.2.3",
-    "load-grunt-tasks": "^3.2.0"
+    "load-grunt-config": "^0.17.2"
   }
 }
