+++ tasks/open.js
@@ -1,0 +1,7 @@
+// Opens the web server in the browser
+// https://github.com/jsoverson/grunt-open
+module.exports = {
+  server: {
+    path: 'http://localhost:<%= port %>'
+  }
+};
